Create README.md
Are you passionate about harnessing the power of data to drive innovation and solve complex problems? Do you thrive in a dynamic, collaborative environment where your skills in data science, engineering, and machine learning can make a real impact?

Requirements:
	•	Strong background in data science, machine learning, and statistics.
	•	Proficiency in programming languages such as Python, R, or Java.
	•	Experience with data engineering tools and technologies (e.g., SQL, Hadoop, Spark).
	•	Excellent problem-solving and communication skills.
If you're ready to take your career to the next level and work on cutting-edge projects that shape the future, we want to hear from you! Apply now to join our dynamic team and be part of our journey towards innovation and excellence.

we are changing the world and we are looking for more people to get this done.

If you are interested please DM me or send me your resume, and GitHub to gabantsil@gmail.com


*Job Title: Data Scientist - Advanced Statistics & Machine Learning*

*Job Description:*

We're looking for a Data Scientist to help us make sense of bog amounts of diverse web data. Your main tasks involve scraping, cleaning, and analyzing data, applying statistical methods and machine learning to uncover meaningful insights.

*Key Responsibilities:*

1. Extract data from various online sources using web scraping techniques.
2. Clean and preprocess data for accuracy and reliability.
3. Apply statistical methods and machine learning to find patterns and insights.
4. Improve existing models including embeddings, svm and nl. reg. models
5. Develop and implement data models for predictive analytics.
6. Use Python for data manipulation, analysis, and visualization.
7. Navigate web structures with HTML and CSS skills.
8. Develop, Test and Deploy Models on state of the Art ML Pipelines 

*Qualifications:*

1. Bachelor's degree or higher in a quantitative field.
2. Experience in web scraping, data cleaning, and statistical analysis.
3. Strong Python skills for efficient coding.
4. Familiarity with HTML and CSS for web structure understanding.
5. Solid understanding of mathematical concepts for data analysis.
6. Clear english communication skills for conveying complex findings.
7. Problem-solving attitude, no fear to make and discuss mistakes, curiosity to learn and ability to collaborate in a team.

*Why Join Us:*

- Work with diverse datasets and cutting-edge technologies.
- Join a collaborative environment that values innovation.
- Continuous learning and professional development opportunities.
- Remote Work
- Competitive salary 
- Midsize and enterprise customer and startup mentality 
- Chance to become a Team Lead